+++ scrapy_cffi/templates/spiders/settings.py
@@ -55,6 +55,7 @@ def create_settings(spider_path, env_path=None, used_redis=False, used_rabbitmq=
     elif used_redis:
         settings.SCHEDULER = "scrapy_cffi.scheduler.RedisScheduler" # Starting the Redis scheduler requires configuring Redis information
         settings.REDIS_INFO.URL = "redis://127.0.0.1:6379"
+        settings.SCHEDULER_PERSIST = False
         # Optional: shared Redis Stream consumer-group defaults for RedisSpider (spider attrs override)
         # from scrapy_cffi.models import RedisStreamConsumerInfo, RedisIngressMode
         # settings.REDIS_STREAM_INFO = RedisStreamConsumerInfo(